Replacing `CitationAction` with `TestAction`. Removing `/citations` directory, and `Tab`, `TabPanel`, and `Tabs` components.
