+++ src/modules/lists/components/CitationAction/styles.css
@@ -1,6 +1,21 @@
-.copy-citation {
-  border: solid 1px var(--search-color-grey-400);
-  box-shadow: none;
-  max-height: 40vh;
-  overflow-y: auto;
+.tabs {
+  border-bottom: 2px solid var(--search-color-grey-400);
+  display: flex;
+}
+
+.tabs__tab {
+  background: none;
+  border: 0 solid transparent;
+  border-radius: 4px 4px 0 0;
+  border-width: 3px 1px;
+  margin-bottom: -2px;
+  padding: .5rem 1.25rem;
+}
+
+.tabs__tab--active {
+  border-color: var(--search-color-grey-400);
+  border-bottom-color: var(--search-color-grey-100);
+  border-top-color: currentColor;
+  color: var(--search-color-blue-400);
+  font-weight: 600;
 }
