--- src/modules/lists/components/TestAction/styles.css
@@ -1,21 +1,0 @@
-.tabs {
-  border-bottom: 2px solid var(--search-color-grey-400);
-  display: flex;
-}
-
-.tabs__tab {
-  background: none;
-  border: 0 solid transparent;
-  border-radius: 4px 4px 0 0;
-  border-width: 3px 1px;
-  margin-bottom: -2px;
-  padding: .5rem 1.25rem;
-}
-
-.tabs__tab--active {
-  border-color: var(--search-color-grey-400);
-  border-bottom-color: var(--search-color-grey-100);
-  border-top-color: currentColor;
-  color: var(--search-color-blue-400);
-  font-weight: 600;
-}
