--- src/modules/reusable/components/Tabs/styles.css
@@ -1,59 +1,0 @@
-div[role="tablist"] {
-  display: flex;
-  flex-direction: column;
-}
-
-@media only screen and (min-width: 641px) {
-  div[role="tablist"] {
-    align-items: flex-end;
-    flex-direction: row;
-  }
-}
-
-div[role="tablist"] > * {
-  all: unset;
-  border-color: var(--search-color-grey-400);
-  border-style: solid;
-  border-width: 0;
-  border-left-width: 3px;
-}
-
-@media only screen and (min-width: 641px) {
-  div[role="tablist"] > * {
-    border-bottom-width: 2px;
-    border-left-width: 0;
-    flex: 0 0 auto;
-  }
-
-  div[role="tablist"] > div:empty {
-    flex: 1 1 auto;
-  }
-}
-
-button[role="tab"] {
-  cursor: pointer;
-  padding: 0.5rem 1.25rem;
-}
-
-button[role="tab"]:hover {
-  text-decoration: underline;
-}
-
-button[role="tab"][aria-selected="true"] {
-  background: var(--search-color-grey-200);
-  border-left-color: currentColor;
-  color: var(--search-color-blue-400);
-  font-weight: 600;
-}
-
-@media only screen and (min-width: 641px) {
-  button[role="tab"][aria-selected="true"] {
-    background: white;
-    border-bottom-color: transparent;
-    border-left-color: var(--search-color-grey-400);
-    border-left-width: 1px;
-    border-right-width: 1px;
-    border-top-color: currentColor;
-    border-top-width: 3px;
-  }
-}
