+++ src/modules/reusable/index.js
@@ -8,9 +8,6 @@ import Dialog from './components/Dialog';
 import H1 from './components/H1';
 import Icon from './components/Icon';
 import ImagePlaceholder from './components/ImagePlaceholder';
-import Tab from './components/Tab';
-import TabPanel from './components/TabPanel';
-import Tabs from './components/Tabs';
 import TruncateText from './components/TruncateText';
 import useWindowWidth from './components/useWindowWidth';
 
@@ -28,9 +25,6 @@ export {
   H1,
   Icon,
   ImagePlaceholder,
-  Tab,
-  Tabs,
-  TabPanel,
   TruncateText,
   useWindowWidth
 };
